LPS-54223 move portlet statistics to module
